Quadratic Describer
Utilizes the quadratic formula. After asking for the user input for the "a", "b", and "c" values of the formula, it will use it to calculate solutions for a quadratic equation. Afterwards, it provides a description of the result, such as where the graph opens up, the axis of symmetry, vertex, etc. At the end, it will ask if the user wants to keep going. If the user types "quit", the program will stop.