Guessing Game
Asks a user to guess a randomly generated integer between a provided range. First asks user for their name and greets them with it. Then, it asks for the lowest number and highest number as a range. While the user does not guess correctly, it will say whether the guess is too high or low.