The Hobbit Revised
Practice assignment: A class to practice inheritance and Array Lists by making Dwarf and Wizard subclasses of the superclass Traveler. In the client class "ThereAndBackAgain", it creates an Array List as a party for different Travelers to store. There are two additional methods that create a party and allows everyone in a party to travel the same amount of miles.