Survival Island Game
This project is a story game that allows a user to interact with objects while progressing through a story. It begins with the user choosing a character: a rabbit or a dinosaur. The first scene is placed in a desert, and the only object available to click is a palm tree. Upon click, a coconut and treasure chest will drop down, and clicking the treasure chest will advance them to a campsite. To survive the night, they must light up a bonfire and eat a coconut to satisfy their hunger. Once they do so, they will sleep through the night and later walk into a grassland with a cove. Here, the user will meet a bear that will ask them how they found their territory. If the user says the wrong thing throughout the conversation, they will be sent back to the beginning of the scene until they make the correct choice. Once they are done, the bear will offer to sell materials, and the user must purchase ones that can be used to build a boat and escape the island. After the task is complete, they will leave the island and an ending screen will show. 
I wrote this in the context of my Computer Science Principles class' final project, and it demonstrates my skills of game concept designs, loops, familiarity of Snap!, conditional statements, and storing objects.