The Hobbit Revised
Written as a practice assignment for my AP Computer Science class, this project demonstrates my skills of inheritance, polymorphism, ArrayList, and one-dimensional arrays. The file contains a client class "ThereAndBackAgain", a super class "Traveler", and three subclasses of the Traveler "Hobbit", "Dwarf", and "Wizard". The subclasses inherits the travel, getDistanceTraveled, and getName methods from the superclass of "Traveler", but each of them will modify the methods. For example, a wizard will travel 3 times the original amount in the travel method and their name will be in the format of "[Wizard] the [Color]". Others, however, may not be modified, such as "Dwarf" and "Hobbit". 
In the client class, three different types of Travelers will be created and will be added into a one-dimensional array labeled as "party". Everyone in the "party" will travel 50 miles, and the class will print out the distance traveled by each traveler.
The class also contains two static methods: createParty and allTravel. The createParty method takes in an ArrayList of Traveler type and String array of names. Using a for-each loop, it will create a Traveler with a name from the array, and add the Traveler to the ArrayList parameter. The second method allTravel takes in an ArrayList of Traveler type and an int number of miles. Using a for-each loop, each traveler will travel an int number of miles. 
In the second part of the client class, a one-dimensional array "dwarfNames" is initialized with a list of String names. A new ArrayList party labeled as "party2" will be created, and a hobbit and a wizard will be added to it. Using the createParty method, it will put party2 and dwarfNames will be put into the parameter. It will also call the allTravel method and put in party2 and 100 as the parameters. Using a for each loop, the class will print out how many miles each Traveler in the party has traveled.