Quadratic Describer
Written as an assignment for my AP Computer Science class, this project calculates a quadratic equation using a user's inputs of the "a", "b", and "c" values of the standard form of the quadratic equation. After it calculates the solutions, it will provide a description of the equation, such as where the graph opens up, the axis of symmetry, vertex, etc. At the end, it will ask if the user wants to keep going. If the user types "quit", the program will stop, and otherwise, it will keep asking the user for the "a", "b", and "c" inputs. This demonstrates my skills of utilizing while loops and basic mathematical skills in a programming setting.