Text Excel
Written as a project for my AP Computer Science class, this project is a miniature version of the Microsoft Excel, creating a 12x20 spreadsheet grid that displays everything a user puts in. The columns are labeled with letters from A-L, and rows are labeled from 1-20. This demonstrates my skills of inheritance, polymorphism, arrays, parsing, conditional statements, and loops. It contains three interfaces (Grid, Cell, Location), five superclasses that implement the interfaces (EmptyCell, TextCell, RealCell, Spreadsheet, and SpreadsheetLocation), three subclasses (FormulaCell, PercentCell, ValueCell), and a client class "TextExcel". 
When the program is ran, it will ask a user to enter a command. Here, the user has four options: assignment, inspection, clear, or clear all. In assignment, a user can call a cell and set it equal to a String (calls the TextCell), a numeric value (calls RealCell or ValueCell), a percent (calls PercentCell), or a formula (calls FormulaCell). With this, the spreadsheet stores the cell in a 2D array in respect to the location that the user referenced. The user can continue to assign and replace different cells. In inspection, a user can type in a cell location, and the program will return the value stored in the cell, and if the cell has nothing, the program will give an error. In clear, a user can type clear and a cell location to erase anything in that location and anything stored in the 2D array spreadsheet. In clear all, everything on the spreadsheet and in the 2D array spreadsheet will be erased. This project will keep running unless a user types something in an incorrect format or a function that may crash the program.