Country Data
Written as a project for my AP Computer Science class, this project scans csv files of countries and their associated numeric values over a period of time, and then it displays datasets and summaries for each country. For example, we have data of the percent of population that has access to electricity for the countries Angola, Bangladesh, Benin, and France from 2015-2020. The program will then take in the the file and display it in a table-like format; a line of years spaced out will be displayed first, a line of the corresponding data below the years will be displayed second, and then a sentence describing the dataset will be displayed, such as "This is the "Access of Electricity (% of population) for Angola". After these, the minimum value, maximum value, and type of trend will show below. This will repeat for each country.