Guessing Game
Written as an assignment for my AP Computer Science class, this project is a game where it repeatedly asks a user for their guess until the correct number is guessed. It demonstrates my skills of using loops, scanners, conditional statements, and throwing arguments. When the program is first ran, it will ask the user for their name and then greet them with it. Afterwards, it asks for a small number to be the minimum for a range and a higher number as the maximum for the range. If the range is in a wrong format (the lower number is higher than the high number), an argument will be thrown. The program will generate a random integer between the range, and then ask the user to guess. For each incorrect guess, the program will give a hint--whether the number is higher or lower than the guess. When the user guesses correctly, the program will say, "You guessed right!" and follow with what the number was. At the end, it will ask if the user would like to play again. If the user says yes, the process will repeat; otherwise, it will thank the user for playing. 
Asks a user to guess a randomly generated integer between a provided range. First asks user for their name and greets them with it. Then, it asks for the lowest number and highest number as a range. While the user does not guess correctly, it will say whether the guess is too high or low.